feat(cli,runtime): a cross-version resume is judged (F-P21 · NEP-0014)
A resume under an engine different from the recording one is now
judged BEFORE the fold, never assumed: the refusal names both
versions and the exact --resume-compat <VERSION> teaching, a wrong
token is its own named refusal, and a pre-versioning trace rides
the unrecorded token (the silent cross-version degradation dies).
The declared compat discharges the crossing and is ATTESTED — the
run's boot manifest journals resumed_from_engine plus
resume_compat: declared, and the evidence pack projects both under
engine.

The judgment lives in nika-dap (judge_version · judge_resume →
CompatVerdict Proceed/Refuse · one source); the CLI maps Proceed
to the fold and Refuse to the environment class; the runtime only
stamps the attestation the composer already judged. An exact match
proceeds without a claim, and the keyless-trace honest degradation
(ADR-099) is unchanged — the law judges the VERSION, not the keys.

Fixtures (unit + binary plane): exact resumes silently · mismatch
refuses naming 0.0.0-test x this engine with the exact teaching ·
the declared compat resumes and the new journal attests the
crossing · a wrong token refuses named · a versionless trace
refuses with the unrecorded token, then proceeds declared.

The lot-3 ratchet folds ride here: check() maps each analyzer error
through conformance_violation (fn length) · the prologue's NEP-0014
fields extract to nep_0014_fields (fn length) · the evidence-pack
tests split to evidence/tests.rs (the 1,500-LOC ceiling).
